+++ src/main/kotlin/de/pause/features/shop/data/repo/DishRepository.kt
@@ -4,7 +4,7 @@ package de.pause.features.shop.data.repo
 import de.pause.database.suspendTransaction
 import de.pause.features.shop.data.dao.Dish
 import de.pause.features.shop.data.dao.DishTable
-import de.pause.model.DishDto
+import de.pause.features.shop.data.dto.DishDto
 import org.jetbrains.exposed.sql.SqlExpressionBuilder.eq
 import org.jetbrains.exposed.sql.deleteWhere
 import org.joda.time.DateTime
