+++ src/main/kotlin/de/pause/plugins/Security.kt
@@ -3,7 +3,7 @@ package de.pause.plugins
 import com.auth0.jwt.JWT
 import com.auth0.jwt.algorithms.Algorithm
 import de.pause.features.user.data.dto.UserPrincipal
-import de.pause.model.UserRole
+import de.pause.util.UserRole
 import io.ktor.server.application.*
 import io.ktor.server.auth.*
 import io.ktor.server.auth.jwt.*
