+++ src/main/kotlin/de/pause/util/Constraints.kt
@@ -1,4 +1,4 @@
-package de.pause.model
+package de.pause.util
 
 object Constraints {
     const val VALID_USER_EMAIL_SUFFIX = "email.com"
