+++ src/test/kotlin/de/pause/SecurityTests.kt
@@ -3,7 +3,7 @@ package de.pause
 import com.auth0.jwt.JWT
 import com.auth0.jwt.algorithms.Algorithm
 import de.pause.features.user.data.dto.LoginRequest
-import de.pause.model.UserRole
+import de.pause.util.UserRole
 import io.ktor.client.plugins.contentnegotiation.*
 import io.ktor.client.request.*
 import io.ktor.client.statement.*
